Export only one query diff result component
